+++ fbw-a380x/src/systems/instruments/src/MsfsAvionicsCommon/EcamMessages/AbnormalSensed/ata29-30.tsx
@@ -11,7 +11,8 @@ import { AbnormalProcedure, ChecklistLineStyle } from 'instruments/src/MsfsAvion
 //    1 for normal checklists,
 //    2 for infos,
 //    3 for INOP SYS,
-//    4 for limitations (not populated yet here),
+//    4 for limitations,
+//    7 for deferred procedures,
 //    8 for ABN sensed procedures,
 //    9 for ABN non-sensed procedures
 
@@ -235,6 +236,7 @@ export const EcamAbnormalSensedAta2930: { [n: number]: AbnormalProcedure } = {
       {
         name: 'IF G SYS PRESSURE FLUCTUATES',
         sensed: false,
+        condition: true,
       },
       {
         name: 'G ENG 1 PMP A AND B',
@@ -264,6 +266,7 @@ export const EcamAbnormalSensedAta2930: { [n: number]: AbnormalProcedure } = {
       {
         name: 'IF Y SYS PRESSURE FLUCTUATES',
         sensed: false,
+        condition: true,
       },
       {
         name: 'Y ENG 3 PMP A AND B',
@@ -287,7 +290,7 @@ export const EcamAbnormalSensedAta2930: { [n: number]: AbnormalProcedure } = {
     ],
   },
   290800021: {
-    title: '\x1b<4m\x1b4mHYD\x1bm G  RSVR LEVEL LO',
+    title: '\x1b<4m\x1b4mHYD\x1bm G RSVR LEVEL LO',
     sensed: true,
     items: [
       {
@@ -320,7 +323,7 @@ export const EcamAbnormalSensedAta2930: { [n: number]: AbnormalProcedure } = {
     ],
   },
   290800022: {
-    title: '\x1b<4m\x1b4mHYD\x1bm Y  RSVR LEVEL LO',
+    title: '\x1b<4m\x1b4mHYD\x1bm Y RSVR LEVEL LO',
     sensed: true,
     items: [
       {
@@ -373,12 +376,12 @@ export const EcamAbnormalSensedAta2930: { [n: number]: AbnormalProcedure } = {
     items: [],
   },
   290800027: {
-    title: '\x1b<4m\x1b4mHYD\x1bm G  SYS COOLING FAULT',
+    title: '\x1b<4m\x1b4mHYD\x1bm G SYS COOLING FAULT',
     sensed: true,
     items: [],
   },
   290800028: {
-    title: '\x1b<4m\x1b4mHYD\x1bm Y  SYS COOLING FAULT',
+    title: '\x1b<4m\x1b4mHYD\x1bm Y SYS COOLING FAULT',
     sensed: true,
     items: [],
   },
@@ -467,7 +470,7 @@ export const EcamAbnormalSensedAta2930: { [n: number]: AbnormalProcedure } = {
     items: [],
   },
   290800035: {
-    title: '\x1b<4m\x1b4mHYD\x1bm G SYS PRESS LO',
+    title: "\x1b<4m\x1b4mHYD\x1bm \x1b'mG SYS PRESS LO\x1bm",
     sensed: true,
     items: [
       {
@@ -508,7 +511,7 @@ export const EcamAbnormalSensedAta2930: { [n: number]: AbnormalProcedure } = {
     ],
   },
   290800036: {
-    title: '\x1b<4m\x1b4mHYD\x1bm Y SYS PRESS LO',
+    title: "\x1b<4m\x1b4mHYD\x1bm \x1b'mY SYS PRESS LO\x1bm",
     sensed: true,
     items: [
       {
@@ -604,7 +607,7 @@ export const EcamAbnormalSensedAta2930: { [n: number]: AbnormalProcedure } = {
     ],
   },
   290800039: {
-    title: '\x1b<4m\x1b4mHYD\x1bm G+Y SYS PRESS LO',
+    title: "\x1b<4m\x1b4mHYD\x1bm \x1b'mG+Y SYS PRESS LO\x1bm",
     sensed: true,
     items: [
       {